# stockbuddy-backend-main/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StockBuddy API",
    "description": "Comprehensive API for StockBuddy - Inventory, Transaction, and Event Management System with multi-tenant architecture, JWT authentication, and extensive business logic for managing members, vendors, banks, transactions, inventory records, events, and more.",
    "version": "1.0.0",
    "contact": {
      "name": "StockBuddy API Support",
      "email": "support@stockbuddy.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base Path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint. Include as: Authorization: Bearer <token>"
      }
    },
    "schemas": {
...
      "Organization": {
        "type": "object",
        "description": "Organization/company information",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique organization identifier"
          },
          "name": {
            "type": "string",
            "description": "Organization display name"
          },
          "owner_user_id": {
            "type": "string",
            "description": "ID of organization owner"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Organization creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
      "Member": {
        "type": "object",
        "description": "Member/customer record with team affiliation and membership details",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique member identifier"
          },
          "name": {
            "type": "string",
            "description": "Member's full name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Member's email address"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BANNED"
            ],
            "description": "Member account status"
          },
          "team_name": {
            "type": "string",
            "description": "Team display name"
          },
          "membership_price": {
            "type": "number",
            "description": "Membership fee amount"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/organizations": {
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Create Organization",
        "description": "Create a new organization and assign the current user as owner. User must not already belong to an organization.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "example": "Acme Corporation"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "organization": {
                      "$ref": "#/components/schemas/Organization"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/members": {
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "List Members",
        "description": "Get filtered list of organization members. Supports filtering by team, fixture, game, and search terms. Only returns ACTIVE members unless specifically filtered.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search term for member name or email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team_names",
            "in": "query",
            "description": "Comma-separated list of team names to filter by",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Members retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Member"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Members"
        ],
        "summary": "Create Member",
        "description": "Create a new member record. Team information will be resolved via API Football if team_name is provided.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "team_name",
                  "membership_price"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john.doe@example.com"
                  },
                  "team_name": {
                    "type": "string",
                    "example": "Arsenal"
                  },
                  "membership_price": {
                    "type": "number",
                    "example": 50
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Member created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Member"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}